Move `arb_prime` to `prime::tests` module
