+++ ext/crates/fp/src/vector/mod.rs
@@ -11,8 +11,6 @@ pub use inner::*;
 
 #[cfg(test)]
 mod tests {
-    use std::sync::OnceLock;
-
     use itertools::Itertools;
     use proptest::prelude::*;
     use rstest::rstest;
@@ -21,7 +19,7 @@ mod tests {
     use crate::{
         field::{field_internal::FieldInternal, fp::F2, Fp},
         limb,
-        prime::{Prime, ValidPrime},
+        prime::{tests::arb_prime, Prime, ValidPrime},
     };
 
     pub struct VectorDiffEntry {
@@ -97,32 +95,6 @@ mod tests {
     fn random_vector(p: u32, dimension: usize) -> Vec<u32> {
         let mut rng = rand::thread_rng();
         (0..dimension).map(|_| rng.gen_range(0..p)).collect()
-    }
-
-    /// An arbitrary `ValidPrime` in the range `2..(1 << 24)`, plus the largest prime that we support.
-    fn arb_prime() -> impl Strategy<Value = ValidPrime> {
-        static TEST_PRIMES: OnceLock<Vec<ValidPrime>> = OnceLock::new();
-        let test_primes = TEST_PRIMES.get_or_init(|| {
-            // Sieve of erathosthenes
-            const MAX: usize = 1 << 24;
-            let mut is_prime = Vec::new();
-            is_prime.resize_with(MAX, || true);
-            is_prime[0] = false;
-            is_prime[1] = false;
-            for i in 2..MAX {
-                if is_prime[i] {
-                    for j in ((2 * i)..MAX).step_by(i) {
-                        is_prime[j] = false;
-                    }
-                }
-            }
-            (0..MAX)
-                .filter(|&i| is_prime[i])
-                .map(|p| ValidPrime::new_unchecked(p as u32))
-                .chain(std::iter::once(ValidPrime::new_unchecked(2147483647)))
-                .collect()
-        });
-        (0..test_primes.len()).prop_map(|i| test_primes[i])
     }
 
     /// An arbitrary (prime, dimension) pair
